changes: merge code to ApiHealthCheckUtils , update the date issue in report ,removed print statement

+++ src/main/java/com/gemini/quartzReporting/TestCase_Details.java
@@ -59,11 +59,11 @@ class TestCase_Details {
 		return tc_run_id;
 	}
 
-	public float getStart_time() {
+	public long getStart_time() {
 		return start_time;
 	}
 
-	public float getEnd_time() {
+	public long getEnd_time() {
 		return end_time;
 	}
